requirements.txt
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
filetype==1.0.7
pdfminer.six==20201018
pycparser==2.20
six==1.15.0
sortedcontainers==2.3.0
